combine scheduled_queries
